_Projects/Deep_Research_Agent/pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "It will give you the reply after deep Research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.8",
    "python-dotenv>=1.0.1",
    "openai-agents>=0.2.5",
    "tavily-python>=0.7.10",
    "openai==1.98.0",
    "mem0ai>=0.1.116",
]
